requirements.txt
fastapi
uvicorn[standard]
python-dotenv
torch
torchvision
transformers
scikit-learn
pillow
pydantic
faiss-cpu
python-multipart
numpy


